Move sequence plannode to presto-main
Moved Sequence node to presto main since a dependency of com.google.common.graph was needed which is not available in spi and previous disussions have suggested not adding that dependency to spi
