+++ presto-main/src/main/java/com/facebook/presto/sql/planner/plan/SequenceNode.java
@@ -11,9 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spi.plan;
+package com.facebook.presto.sql.planner.plan;
 
 import com.facebook.presto.spi.SourceLocation;
+import com.facebook.presto.spi.plan.PlanNode;
+import com.facebook.presto.spi.plan.PlanNodeId;
 import com.facebook.presto.spi.relation.VariableReferenceExpression;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -23,7 +25,7 @@ import java.util.List;
 import java.util.Optional;
 
 public class SequenceNode
-        extends PlanNode
+        extends InternalPlanNode
 {
     // cteProducers {l1,l2,l3} will be executed in {l3, l2,l1} order
     private final List<PlanNode> cteProducers;
@@ -89,7 +91,7 @@ public class SequenceNode
     }
 
     @Override
-    public <R, C> R accept(PlanVisitor<R, C> visitor, C context)
+    public <R, C> R accept(InternalPlanVisitor<R, C> visitor, C context)
     {
         return visitor.visitSequence(this, context);
     }
