+++ presto-spi/src/main/java/com/facebook/presto/spi/plan/PlanVisitor.java
@@ -100,11 +100,6 @@ public abstract class PlanVisitor<R, C>
         return visitPlan(node, context);
     }
 
-    public R visitSequence(SequenceNode node, C context)
-    {
-        return visitPlan(node, context);
-    }
-
     public R visitConnectorJoinNode(ConnectorJoinNode node, C context)
     {
         return visitPlan(node, context);
